libcare-dump: add libcare-dump auxiliary tools
Sometimes we need a auxiliary tools to get kpatch information for
analysis, such as: whether the assembly code of the kpatch meets our
expectiations, what sysmbols are included by the kpatch, why the kpatch
did not take effect and so on.

So we provide a libcare-dump tools to help analyze the content of the
kpatch, it can provide the following functions:
    1. convert kpatch file to elf file, so we can use readelf and
       objdump tools to analyze it;
    2. print the kpatch header related information(to be realized);

The libcare-dump tools use like this:
    1. libcare-dump -c -i <input> -o <output>;

For detailed usage, please using: libcare-dump --help.

Signed-off-by: Bihong Yu <yubihong@huawei.com>
Signed-off-by: imxcc <xingchaochao@huawei.com>