process: add a flag to mark unpatch target elf
When we unpatch a process that multiple object has been patched, libcare
will try to unpatch all patch in function kpatch_unapply_patches() which
is not what we expected. In order to slove the problem, we add the
is_unpatch_target_elf flag to mark the unpatch target elf object, and
check the flags before unpatch for object who have been patched.

Signed-off-by: Bihong Yu <yubihong@huawei.com>
Signed-off-by: imxcc <xingchaochao@huawei.com>