strip: settle libcare-dump output elf file can not be objdump bug
Fix the problem that the output elf file of the libcare-dump can't be
objdump.

Objdump's bfd_check_format_matches() call bfd_elf64_object_p() to check
elf format, it will check whether the sh_link and the sh_info papameter
is legal. We only change the sh_type to SHT_NOBITS of elf section
header(GElf_Shdr), if the section we don't need. Then we use
/usr/bin/strip to strip unneed section, this will no change the section
of sh_type=SHT_NOBITS, resulting in wrong sh_link and sh_info value.

To fix this problem, we change the sh_link and the sh_info to 0 together
if the section we don't need.

Signed-off-by: Bihong Yu <yubihong@huawei.com>
Signed-off-by: imxcc <xingchaochao@huawei.com>