process: adapt to recognize kpatch region
Now the kpatch region consists of two part mem region. The first part is
kpatch header + jmp table + .kpatch.text and the access permision is r-xp,
the second part is .kpatch.data + rela + user_undo and the access
permison is rw-p.

We just need to use the first part to recognize and get the kpatch info.

Signed-off-by: Bihong Yu <yubihong@huawei.com>
Signed-off-by: imxcc <xingchaochao@huawei.com>