kpatch_gensrc.c: support ignoring functions which we don't need
Adding or deleting statements in the function will cause the
subsequent __LINE__ macro to be offset. The change will also be
recognized by the patch tools and made into the patch file,
which is not what we expected.

So we add a configure file 'patch_func_names.conf'. We can write
the functions we need into the configuration file. When the patch
tools make a patch, they will filter out other functions that we
don't need based on the content in the configuration file.

Signed-off-by: Jiajie Li <lijiajie11@huawei.com>
Signed-off-by: Bihong Yu <yubihong@huawei.com>
Reviewed-by: Bihong Yu <yubihong@huawei.com>